Refactor: streamline method replacement
This makes the main replaceGlobalMethodCalls function read more easily from top to bottom, with the various cases being broken down in a linear fashion. This comes at the cost of duplicating some work into the helper functions isPropertyOfConstructor and isPropertyOfNamespace, by for example not sharing the baseConstructSignature extraction code. But the clarity improvement is worth it.