Improve detection of data properties
Closes #33. Closes #56 by superceding it.

The tests for someFn.{length,prototype,name} revealed that we were not handling data property detection well when the base type was a non-built-in function. This restructuring makes the data property detection more robust, which also allows us to use it more intuitively, earlier in the process. This commit also removes the usage of isDataProperty in isWindowGlobal while trying to eliminate property access. Instead we can just eliminate any property accesses.